Cast id col to Int64 in ancestor_id fill_null for safety

+++ hstrat/_auxiliary_lib/_alifestd_try_add_ancestor_id_col_polars.py
@@ -33,7 +33,7 @@ def alifestd_try_add_ancestor_id_col_polars(
         ancestor_id=pl.col("ancestor_list")
         .str.extract(r"(\d+)", 1)
         .cast(pl.Int64, strict=False)
-        .fill_null(pl.col("id"))
+        .fill_null(pl.col("id").cast(pl.Int64))
     )
